Update Mark Color Matt Chart and Table

+++ WebApplication1/WebApplication1/Models/FTWip.cs
@@ -16,7 +16,7 @@ namespace WebApplication1.Models
         public string FTDevice { get; set; }
         public int Sequence { get; set; }
         public string S_Color { get; set; }
-        public int Qty_Production { get; set; }
+        public decimal Qty_Production { get; set; }
         public int Kpcs { get; set; }
         public LotState Lot_State { get; set; }
         public string MachineWip { get; set; }
